refactor: move blog theme component type defs to theme-classic
